+++ ds-conta-corrente/src/main/java/com/farias/banco/dscontacorrente/App.java
@@ -4,13 +4,11 @@ import org.springframework.amqp.rabbit.annotation.EnableRabbit;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.cloud.stream.annotation.EnableBinding;
 
-import com.farias.banco.dscontacorrente.config.app.AppConfig;
-import com.farias.banco.dscontacorrente.config.broker.BrokerInput;
+import com.farias.banco.dscontacorrente.modules.config.app.AppConfig;
+
 @EnableRabbit
 @EnableEurekaClient
 @EnableFeignClients
