--- ds-conta-corrente/src/main/java/com/farias/banco/dscontacorrente/config/app/AppConfig.java
@@ -1,21 +1,0 @@
-package com.farias.banco.dscontacorrente.config.app;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-import lombok.Data;
-
-@ConfigurationProperties("banco")
-public class AppConfig {
-	
-	private final Agencia agencia = new Agencia();
-	
-	public Agencia getAgencia() {
-		return agencia;
-	}
-	
-	@Data
-	public static class Agencia {
-		private Integer numero;
-	}
-
-}
