+++ ds-conta-corrente/src/main/java/com/farias/banco/dscontacorrente/modules/config/feing/FeignConfig.java
@@ -1,0 +1,36 @@
+package com.farias.banco.dscontacorrente.modules.config.feing;
+
+import org.springframework.beans.factory.ObjectFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.http.HttpMessageConverters;
+import org.springframework.cloud.openfeign.support.PageJacksonModule;
+import org.springframework.cloud.openfeign.support.SpringEncoder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.farias.banco.dscontacorrente.modules.integration.feign.impl.CustomJacksonModule;
+import com.fasterxml.jackson.databind.Module;
+
+import feign.codec.Encoder;
+
+@Configuration
+public class FeignConfig {
+
+	@Autowired
+	private ObjectFactory<HttpMessageConverters> messageConverters;
+
+	@Bean
+	public Encoder feignEncoder() {
+		return new PageableQueryEncoder(new SpringEncoder(messageConverters));
+	}
+	
+	@Bean
+	public Module pageJacksonModule() {
+		return new PageJacksonModule();
+	}
+
+	@Bean
+    public Module myJacksonModule() {
+        return new CustomJacksonModule();
+    }
+}
