+++ ds-conta-corrente/src/main/java/com/farias/banco/dscontacorrente/utils/mapper/ContaCorrenteMapper.java
@@ -3,8 +3,8 @@ package com.farias.banco.dscontacorrente.utils.mapper;
 import org.mapstruct.Mapper;
 
 import com.farias.banco.dscontacorrente.dto.ContaCorrenteResponseDTO;
-import com.farias.banco.dscontacorrente.model.ContaCorrente;
-import com.farias.banco.dscontacorrente.model.Pessoa;
+import com.farias.banco.dscontacorrente.modules.model.ContaCorrente;
+import com.farias.banco.dscontacorrente.modules.model.Pessoa;
 
 @Mapper
 public interface ContaCorrenteMapper {
