+++ ds-conta-corrente/src/test/java/com/farias/banco/dscontacorrente/creators/PessoaCreator.java
@@ -1,0 +1,21 @@
+package com.farias.banco.dscontacorrente.creators;
+
+import com.farias.banco.dscontacorrente.enums.PessoaTipoEnum;
+import com.farias.banco.dscontacorrente.modules.model.Pessoa;
+
+public class PessoaCreator {
+
+	public static Pessoa createPF() {
+		return Pessoa.builder()
+				.id(1l)
+				.cpfCnpj("09152280098")
+				.nome("Jorge paulo")
+				.tipo(PessoaTipoEnum.PF.name())
+				.score(4)
+				.build();
+	}
+
+	public static Pessoa createPJ() {
+		return Pessoa.builder().cpfCnpj("Gomes & Dias - ME").nome("80107417000197").build();	
+	}
+}
